+++ lovely/esdb/tests.py
@@ -89,6 +89,7 @@ def test_suite():
     return unittest.TestSuite((
         create_suite('document/README.rst'),
         create_suite('document/document.rst'),
+        create_suite('document/lazy.rst'),
         create_suite('document/bulk.rst'),
 
         create_suite('properties/property.rst'),